move SelectCountry fragment to a new package countryContacts
