+++ app/src/main/res/layout/fragment_letter.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context="LetterFragment">
+    tools:context=".countryContacts.SelectCountryFragment">
     <ImageView
         android:id="@+id/whales_image"
         android:src="@drawable/country_contacts_whales"
